✈️ refactor error handling
